Turn on Biome's project domain, and break the cycle it finds
The lint that CI runs and the lint that `precommit` runs are already the
same command, and both are already stricter than Biome's default: the
runner in `scripts/biome.ts` treats an info-level diagnostic as a failure,
which `--error-on-warnings` alone does not. What was missing is Biome's
`project` domain, which was off, so every rule that needs to look across
files stayed off with it.

Two of that domain's rules cannot read this project. `noUnresolvedImports`
and `noUndeclaredDependencies` resolve against `node_modules` and a
package manifest, and this repository uses the import map in `deno.json`
with `jsr:` and `npm:` specifiers. With the domain on and both rules left
alone, Biome reports 30,517 problems, which is almost every import in the
tree. Both are off, and the domain is on.

That leaves four findings, and they are one cycle: `listings-edit.ts`
imports `listing-page.ts`, which imports
`listing-page-management-panels.ts`, which imports `listings-edit.ts`
again. The panels file wanted one thing from the far end of that
triangle — `getListingAndGroups`, a loader with two callers and no
opinions. It now lives in `listing-page-data.ts`, which both ends of the
cycle already import, and the triangle is a line.

The lasting gain is the rest of the domain. `noFloatingPromises` and its
siblings report nothing today, and from here on they are enforced.
